python/requirements.txt
scalekit-sdk-python
python-dotenv
requests
anthropic
langchain
langchain-openai
google-adk
litellm
openai
protobuf
crewai
crewai-tools
